Funciona el listado de afiliados Falta la logica de abm de afiliados y de profesionales, con el formulario Compra de bonos -> Falta guardar los cambios

+++ trunk/tp2013/src/ClinicaFrba.Comun/Afiliado.cs
@@ -7,8 +7,7 @@ namespace ClinicaFrba.Comun
 {
     public class Afiliado : User
     {
-        //public static Profile Profile = Profile.Afiliado;
-        public int NroAfiliado { get; set; }
+        public long NroAfiliado { get; set; }
         public EstadoCivil EstadoCivil { get; set; }
         public int CantHijos { get; set; }
         public PlanMedico PlanMedico { get; set; }
